# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EventReach API (mock)",
    "description": "Campaigns and login use the database when DATABASE_URL is set. Everything else is still mock data. Full phone numbers are never returned.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List Campaigns",
        "description": "Your campaigns only. Needs login when the database is on.",
        "operationId": "list_campaigns_campaigns_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Campaign"
                  },
                  "type": "array",
                  "title": "Response List Campaigns Campaigns Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create Campaign",
        "operationId": "create_campaign_campaigns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/campaigns/{campaign_id}/contacts": {
      "get": {
        "tags": [
          "Audience"
        ],
        "summary": "List Contacts",
        "description": "List contacts for a campaign. Requires login when the database is on.",
        "operationId": "list_contacts_campaigns__campaign_id__contacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Language"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segment"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Outcome"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Stage"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/campaigns/{campaign_id}/analytics/funnel": {
      "get": {
        "tags": [
          "Analytics (B1)"
        ],
        "summary": "Funnel",
        "operationId": "funnel_campaigns__campaign_id__analytics_funnel_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Funnel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/r/{token}": {
      "get": {
        "tags": [
          "Registration (public)"
        ],
        "summary": "Registration Page",
        "operationId": "registration_page_r__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/r/{token}/register": {
      "post": {
        "tags": [
          "Registration (public)"
        ],
        "summary": "Register",
        "operationId": "register_r__token__register_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/r/{token}/pay": {
      "post": {
        "tags": [
          "Registration (public)"
        ],
        "summary": "Create Payment",
        "operationId": "create_payment_r__token__pay_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOrder"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/payment": {
      "post": {
        "tags": [
          "Registration (public)"
        ],
        "summary": "Payment Webhook",
        "description": "Gateway callback. Verifies gateway signature when header is present and safely idempotently marks order paid.",
        "operationId": "payment_webhook_webhooks_payment_post",
        "parameters": [
          {
            "name": "x-razorpay-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Razorpay-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentWebhook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Payment Webhook Webhooks Payment Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Campaign": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "template_key": {
            "type": "string",
            "title": "Template Key"
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "languages": {
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "type": "array",
            "title": "Languages"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/Channel"
            },
            "type": "array",
            "title": "Channels"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "contact_count": {
            "type": "integer",
            "title": "Contact Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "template_key",
          "status",
          "languages",
          "channels",
          "contact_count",
          "created_at"
        ],
        "title": "Campaign"
      },
      "CampaignCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "examples": [
              "AI in Healthcare Seminar"
            ]
          },
          "template_key": {
            "type": "string",
            "title": "Template Key",
            "examples": [
              "seminar_invite"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "template_key"
        ],
        "title": "CampaignCreate"
      },
      "CampaignStatus": {
        "type": "string",
        "enum": [
          "draft",
          "details_approved",
          "audience_ready",
          "translations_approved",
          "content_ready",
          "tested",
          "launched",
          "needs_regeneration",
          "ready",
          "running",
          "completed"
        ],
        "title": "CampaignStatus"
      },
      "Channel": {
        "type": "string",
        "enum": [
          "call",
          "sms",
          "email",
          "whatsapp",
          "instagram"
        ],
        "title": "Channel"
      },
      "Contact": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone_masked": {
            "type": "string",
            "title": "Phone Masked",
            "description": "The API never returns a full phone number"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "segment": {
            "type": "string",
            "title": "Segment"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "last_outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "opted_out": {
            "type": "boolean",
            "title": "Opted Out"
          },
          "registration_link": {
            "type": "string",
            "title": "Registration Link"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone_masked",
          "language",
          "segment",
          "stage",
          "last_outcome",
          "attempts",
          "opted_out",
          "registration_link"
        ],
        "title": "Contact"
      },
      "ContactPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Contact"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ContactPage"
      },
      "EventDetails": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "fee_inr": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Fee Inr",
            "default": 0
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "rsvp_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsvp Deadline"
          }
        },
        "type": "object",
        "required": [
          "title",
          "starts_at",
          "venue",
          "city"
        ],
        "title": "EventDetails"
      },
...
      "Funnel": {
        "properties": {
          "contacts": {
            "type": "integer",
            "title": "Contacts"
          },
          "dialed": {
            "type": "integer",
            "title": "Dialed"
          },
          "answered": {
            "type": "integer",
            "title": "Answered"
          },
          "responded": {
            "type": "integer",
            "title": "Responded"
          },
          "confirmed": {
            "type": "integer",
            "title": "Confirmed"
          },
          "registered": {
            "type": "integer",
            "title": "Registered"
          },
          "paid": {
            "type": "integer",
            "title": "Paid"
          }
        },
        "type": "object",
        "required": [
          "contacts",
          "dialed",
          "answered",
          "responded",
          "confirmed",
          "registered",
          "paid"
        ],
        "title": "Funnel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "en",
          "hi",
          "ml",
          "ta"
        ],
        "title": "Language"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "title": "Email",
            "examples": [
              "organizer@example.com"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "Outcome": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "declined",
          "callback",
          "no_answer",
          "voicemail",
          "opted_out",
          "wrong_number",
          "failed"
        ],
        "title": "Outcome"
      },
...
      "PaymentOrder": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "amount_inr": {
            "type": "integer",
            "title": "Amount Inr"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "INR"
          },
          "gateway_key_id": {
            "type": "string",
            "title": "Gateway Key Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "amount_inr",
          "gateway_key_id",
          "status"
        ],
        "title": "PaymentOrder"
      },
      "PaymentWebhook": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "paid"
            ]
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "status"
        ],
        "title": "PaymentWebhook"
      },
...
      "RegisterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "party_size": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Party Size",
            "default": 1
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "description": "Must be true. The person agrees to the privacy notice"
          }
        },
        "type": "object",
        "required": [
          "name",
          "consent"
        ],
        "title": "RegisterRequest"
      },
      "RegisterResult": {
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "requires_payment": {
            "type": "boolean",
            "title": "Requires Payment"
          },
          "amount_inr": {
            "type": "integer",
            "title": "Amount Inr"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "requires_payment",
          "amount_inr"
        ],
        "title": "RegisterResult"
      },
      "RegistrationPage": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "event": {
            "$ref": "#/components/schemas/EventDetails"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "fee_inr": {
            "type": "integer",
            "title": "Fee Inr"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "already_registered": {
            "type": "boolean",
            "title": "Already Registered"
          },
          "event_full": {
            "type": "boolean",
            "title": "Event Full",
            "default": false
          },
          "expired": {
            "type": "boolean",
            "title": "Expired",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "language",
          "event",
          "fee_inr",
          "stage",
          "already_registered"
        ],
        "title": "RegistrationPage"
      },
...
      "Stage": {
        "type": "string",
        "enum": [
          "invited",
          "responded",
          "registered",
          "paid",
          "attended"
        ],
        "title": "Stage"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Paste the token from /auth/login",
        "scheme": "bearer"
      }
    }
  }
}